using chronicle to log events. need to write some tests and make it possible for clients to instantiate billingr with analytics

+++ billingr/src/main/java/com/abatra/billingr/Sku.java
@@ -14,7 +14,7 @@ public interface Sku {
 
     String getCurrency();
 
-    long getPriceAmount();
+    double getPriceAmount();
 
     long getPriceAmountMicros();
